files/requirements.txt
# Core Dependencies
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0

# Video Processing
ffmpeg-python>=0.2.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Machine Learning - CPU
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
onnxruntime>=1.16.0

# Computer Vision & Pose
mediapipe>=0.10.7
scikit-image>=0.21.0

# Data Processing
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0

# Tracking
filterpy>=1.4.5
lap>=0.4.0

# GUI (optional)
PyQt5>=5.15.9
matplotlib>=3.7.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
